+++ src/components/cards/tentacles.tsx
@@ -60,7 +60,7 @@ export const TentacleQuestionComponent = ({
                         onChange={(e) => {
                             const newQuestions = [...$questions];
                             (newQuestions[index].data as typeof data).radius =
-                                parseInt(e.target.value);
+                                parseFloat(e.target.value);
                             questions.set(newQuestions);
                         }}
                     />
